add columns to users table and modify factory a little bit

+++ database/migrations/2020_11_27_221436_create_gifts_table.php
@@ -14,6 +14,7 @@ class CreateGiftsTable extends Migration
     public function up()
     {
         Schema::create('gifts', function (Blueprint $table) {
+          // 商品名、価格、カテゴリコード、画像id、投稿者id、あげたorもらった、相手の性別、相手の年齢、相手との関係コード、シチュエーションコード
           $table->bigIncrements('id');
           $table->string('name', 50);
           $table->integer('price');
@@ -21,7 +22,7 @@ class CreateGiftsTable extends Migration
           $table->integer('img_id');
           $table->integer('post_user_id');
           $table->integer('active_or_passive');
-          $table->integer('opponent_gender');
+          $table->tinyInteger('opponent_gender');
           $table->integer('opponent_age');
           $table->string('relationship_code');
           $table->string('situation_code');
